+++ NEWS.md
@@ -5,7 +5,9 @@ Updated to Julia v0.6 (older versions not supported).
 ##### Changes
 * R logical vectors are converted to `DataVector{Bool}` (instead of `DataVector{Int32}`) [#32]
 * dropped compatibility with Julia versions prior v0.6 [#32]
+* use CodecZlib for gzipped RData files (instead of outdated GZip) [#31]
 
+[#31]: https://github.com/JuliaStats/RData.jl/issues/31
 [#32]: https://github.com/JuliaStats/RData.jl/issues/32
 
 ## RData v0.0.4 Release Notes